moved around files to better fit the flow of components
